chore: add /typography dir with Text compound component
update text path imports
