+++ frontend/src/components/layout/header/HeaderMenu.tsx
@@ -1,11 +1,8 @@
 import { useSwapEvent } from '@/hooks'
 
-import { List } from '@/components/templates'
-import { Text } from '@/components/Text'
-import { LinkEventButton } from '@/components/LinkButton'
+import { List, Text, LinkEventButton, HamburgerDropdownButton } from '@/components'
 import { HeaderMenuList } from './HeaderMenuList'
 import { CloseIcon, HamburgerIcon } from '@/assets'
-import { HamburgerDropdownButton } from '@/components/Button'
 
 import { isString } from '@/utils'
