+++ frontend/src/components/typography/index.ts
@@ -1,0 +1,1 @@
+export * from './Text'
